Split proxy runner into separate inbound and outbound files
